controllers updated

+++ src/controllers/accountsController.js
@@ -16,22 +16,35 @@ const getAccountById = (request, response) => {
     if (error) {
       throw error
     }
-    response.status(200).json(results.rows)
+
+    if(results.rows.length === 0) {
+      response.status(404).send(`there is not account with Id: ${id}`)
+    } else {
+      response.status(200).json(results.rows)
+    }
   })
 }
 
-const createAccount = async (request, response) => {
+const createAccount = (request, response) => {
   const { 
       account_id,
       user_id,
     } = request.body
 
-  await pool.query('INSERT INTO accounts (account_id, user_id) VALUES ($1, $2) RETURNING *', [account_id, user_id], (error, results) => {
-    if (error) {
-      throw error
+  pool.query('SELECT * FROM accounts WHERE account_id = $1', [account_id], (error, results) => {
+
+    if(results.rows.length > 0) {
+      response.status(400).send(`An account exists with account_id: ${account_id}`)
+    } else {
+        pool.query('INSERT INTO accounts (account_id, user_id) VALUES ($1, $2) RETURNING *', [account_id, user_id], (error, results) => {
+          if (error) {
+            throw error
+          }
+          response.status(201).send(`Account successfully added with ID: ${results.rows[0].id}`)
+        })
+      }
     }
-    response.status(201).send(`Account successfully added with ID: ${results.rows[0].id}`)
-  })
+  )
 }
 
 const updateAccount = (request, response) => {
@@ -41,18 +54,26 @@ const updateAccount = (request, response) => {
       user_id,
     } = request.body
 
-  pool.query(
-    'UPDATE accounts SET account_id = $1, user_id = $2 WHERE id = $3', 
-    [
-      account_id, 
-      user_id,
-      id
-    ],
-    (error, results) => {
-      if (error) {
-        throw error
+  pool.query('SELECT * FROM accounts WHERE id = $1', [id], (error, results) => {
+
+    if(results.rows.length === 0) {
+      response.status(404).send(`there is not account with Id: ${id}`)
+    } else {
+        pool.query(
+          'UPDATE accounts SET account_id = $1, user_id = $2 WHERE id = $3', 
+          [
+            account_id, 
+            user_id,
+            id
+          ],
+          (error, results) => {
+            if (error) {
+              throw error
+            }
+            response.status(200).send(`Account modified with ID: ${id}`)
+          }
+        ) 
       }
-      response.status(200).send(`Account modified with ID: ${id}`)
     }
   )
 }
@@ -60,11 +81,19 @@ const updateAccount = (request, response) => {
 const deleteAccount = (request, response) => {
   const id = parseInt(request.params.id)
 
-  pool.query('DELETE FROM Accounts WHERE id = $1', [id], (error, results) => {
-    if (error) {
-      throw error
+  pool.query('SELECT * FROM accounts WHERE id = $1', [id], (error, results) => {
+
+    if(results.rows.length === 0) {
+      response.status(404).send(`there is not account with Id: ${id}`)
+    } else {
+        pool.query('DELETE FROM Accounts WHERE id = $1', [id], (error, results) => {
+          if (error) {
+            throw error
+          }
+          response.status(200).send(`Account deleted with ID: ${id}`)
+        }
+      )
     }
-    response.status(200).send(`Account deleted with ID: ${id}`)
   })
 }
