+++ src/controllers/categoriesController.js
@@ -15,7 +15,11 @@ const getCategoryById = (request, response) => {
     if (error) {
       throw error
     }
-    response.status(200).json(results.rows)
+    if(results.rows.length === 0) {
+      response.send('No category with this id')
+    } else {
+      response.status(200).json(results.rows)
+    }
   })
 }
 
@@ -35,39 +39,56 @@ const createCategory = async (request, response) => {
 }
 
 const updateCategory = (request, response) => {
-  const id = parseInt(request.params.id)
-  const { 
-      category_id,
-      category_name,
-      category_description
-    } = request.body
+  const id = request.params.id
 
-  pool.query(
-    'UPDATE categories SET category_id = $1, category_name = $2, category_description = $3 WHERE id = $4', 
-    [
-      category_id, 
-      category_name,
-      category_description,
-      id
-    ],
-    (error, results) => {
-      if (error) {
-        throw error
-      }
-      response.status(200).send(`Category modified with ID: ${id}`)
+  pool.query('SELECT * FROM categories WHERE id = $1', [id], (error, results) => {
+
+    if(results.rows.length === 0) {
+      response.status(400).send(`no category with Id: ${id}`)
+    } else {
+      const { 
+        category_id,
+        category_name,
+        category_description
+      } = request.body
+
+      pool.query('UPDATE categories SET category_id = $1, category_name = $2, category_description = $3 WHERE id = $4', 
+        [
+          category_id, 
+          category_name,
+          category_description,
+          id
+        ],
+        (error, results) => {
+          if (error) {
+            throw error
+          }
+          response.status(400).send(`category modified with Id: ${id}`)
+        }
+      )
     }
-  )
+  })
 }
 
 const deleteCategory = (request, response) => {
-  const id = parseInt(request.params.id)
+  const id = request.params.id
 
-  pool.query('DELETE FROM categories WHERE id = $1', [id], (error, results) => {
-    if (error) {
-      throw error
+  pool.query('SELECT * FROM categories WHERE id = $1', [id], (error, results) => {
+
+    if(results.rows.length === 0) {
+      response.status(400).send(`no category with Id: ${id}`)
+    } else {
+
+      pool.query('DELETE FROM categories WHERE id = $1', [id], (error, results) => {
+              if (error) {
+                throw error
+              }
+              response.status(400).send(`category deleted with Id: ${id}`)
+            }
+          )
+      }
     }
-    response.status(200).send(`Category deleted with ID: ${id}`)
-  })
+  )
 }
 
 module.exports = {
